making time_t 64 bits

+++ src/main/java/com/frostwire/jlibtorrent/swig/add_torrent_params.java
@@ -208,27 +208,27 @@ public class add_torrent_params {
     return libtorrent_jni.add_torrent_params_seeding_time_get(swigCPtr, this);
   }
 
-  public void setAdded_time(int value) {
+  public void setAdded_time(long value) {
     libtorrent_jni.add_torrent_params_added_time_set(swigCPtr, this, value);
   }
 
-  public int getAdded_time() {
+  public long getAdded_time() {
     return libtorrent_jni.add_torrent_params_added_time_get(swigCPtr, this);
   }
 
-  public void setCompleted_time(int value) {
+  public void setCompleted_time(long value) {
     libtorrent_jni.add_torrent_params_completed_time_set(swigCPtr, this, value);
   }
 
-  public int getCompleted_time() {
+  public long getCompleted_time() {
     return libtorrent_jni.add_torrent_params_completed_time_get(swigCPtr, this);
   }
 
-  public void setLast_seen_complete(int value) {
+  public void setLast_seen_complete(long value) {
     libtorrent_jni.add_torrent_params_last_seen_complete_set(swigCPtr, this, value);
   }
 
-  public int getLast_seen_complete() {
+  public long getLast_seen_complete() {
     return libtorrent_jni.add_torrent_params_last_seen_complete_get(swigCPtr, this);
   }
