+++ src/main/java/com/frostwire/jlibtorrent/swig/file_storage.java
@@ -71,7 +71,7 @@ public class file_storage {
     libtorrent_jni.file_storage_add_file_borrow__SWIG_4(swigCPtr, this, filename, filename_len, path, file_size);
   }
 
-  public void add_file(String path, long file_size, int file_flags, int mtime) {
+  public void add_file(String path, long file_size, int file_flags, long mtime) {
     libtorrent_jni.file_storage_add_file__SWIG_0(swigCPtr, this, path, file_size, file_flags, mtime);
   }
 
@@ -167,7 +167,7 @@ public class file_storage {
     return libtorrent_jni.file_storage_symlink(swigCPtr, this, index);
   }
 
-  public int mtime(int index) {
+  public long mtime(int index) {
     return libtorrent_jni.file_storage_mtime(swigCPtr, this, index);
   }
 
@@ -211,7 +211,7 @@ public class file_storage {
     return libtorrent_jni.file_storage_file_index_at_offset(swigCPtr, this, offset);
   }
 
-  public void add_file(String path, long file_size, int file_flags, int mtime, String symlink_path) {
+  public void add_file(String path, long file_size, int file_flags, long mtime, String symlink_path) {
     libtorrent_jni.file_storage_add_file__SWIG_3(swigCPtr, this, path, file_size, file_flags, mtime, symlink_path);
   }
